+++ semgrep-core/tests/OTHER/rules/metavar_pattern_regex2.js
@@ -1,0 +1,4 @@
+//ruleid:test-mvp-regex-2
+let x = 1
+//ok:test-mvp-regex-2
+let y = "abc"